Rewrite ring buffer handling between bit2float_worker and correlator_node_data_reader_tasklet.
git-svn-id: svn+ssh://jop91.astron.nl/auto/home/jive_cvs/SOFTC/svn/trunk/sfxc@1065 5736e3ea-d426-0410-8aaf-93e336ac3372

+++ include/correlator_node_data_reader_tasklet.h
@@ -56,7 +56,7 @@ private:
   int state;
   enum {IDLE, PROCESSING_STREAM, RECEIVE_DATA};
   /// Indicatates how many bytes still have to be read from the input stream
-  int bytes_left;
+  size_t bytes_left;
 };
 
 #endif // OUTPUT_NODE_DATA_READER_TASKLET_H
